third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-template-areas-get-set-expected.txt

This test checks that grid-template-areas is properly parsed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-template-areas set through CSS.
PASS window.getComputedStyle(gridWithDefaultTemplate).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithSingleStringTemplate).getPropertyValue('grid-template-areas') is "\"area\""
PASS window.getComputedStyle(gridWithTwoColumnsTemplate).getPropertyValue('grid-template-areas') is "\"first second\""
PASS window.getComputedStyle(gridWithTwoRowsTemplate).getPropertyValue('grid-template-areas') is "\"first\" \"second\""
PASS window.getComputedStyle(gridWithSpanningColumnsTemplate).getPropertyValue('grid-template-areas') is "\"span span\""
PASS window.getComputedStyle(gridWithSpanningRowsDotTemplate).getPropertyValue('grid-template-areas') is "\"span\" \".\""
PASS window.getComputedStyle(gridWithDotColumn).getPropertyValue('grid-template-areas') is "\"header .\" \"footer .\""
Test grid-template-areas: initial
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "\"foobar\""
Test grid-template-areas: inherit
PASS window.getComputedStyle(parentElement).getPropertyValue('grid-template-areas') is "\"foo bar\""
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "\"foo bar\""
Test invalid grid-template-areas values.
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"

FIXME: We currently don't validate that the named grid areas are <indent>.
FAIL window.getComputedStyle(element).getPropertyValue('grid-template-areas') should be none. Was "nav-up".
PASS successfullyParsed is true

TEST COMPLETE
